services/training/requirements.txt
scikit-learn==1.4.0
xgboost==2.0.3
mlflow==2.10.2
pandas==2.1.4
psycopg2-binary==2.9.9
boto3==1.34.34
numpy==1.26.3
imbalanced-learn==0.11.0